requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
openai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
